# table: use evalBuffer to improve performance of locatePartition