--- a/table/tables/partition_test.go
+++ b/table/tables/partition_test.go
@@ -278,7 +278,7 @@ func (ts *testSuite) TestGeneratePartitionExpr(c *C) {
 	}
 }
 
-func (ts *testSuite) TestLocateRangePartitionErr(c *C) {
+func (ts *testSuite) TestLocateRangeColumnPartitionErr(c *C) {
 	tk := testkit.NewTestKitWithInit(c, ts.store)
 	tk.MustExec("use test")
 	tk.MustExec(`CREATE TABLE t_month_data_monitor (
@@ -291,6 +291,22 @@ func (ts *testSuite) TestLocateRangePartitionErr(c *C) {
 	)`)
 
 	_, err := tk.Exec("INSERT INTO t_month_data_monitor VALUES (4, '2019-04-04')")
+	c.Assert(table.ErrNoPartitionForGivenValue.Equal(err), IsTrue)
+}
+
+func (ts *testSuite) TestLocateRangePartitionErr(c *C) {
+	tk := testkit.NewTestKitWithInit(c, ts.store)
+	tk.MustExec("use test")
+	tk.MustExec(`CREATE TABLE t_month_data_monitor (
+		id int(20) NOT NULL AUTO_INCREMENT,
+		data_date date NOT NULL,
+		PRIMARY KEY (id, data_date)
+	) PARTITION BY RANGE(id) (
+		PARTITION p0 VALUES LESS THAN (1024),
+		PARTITION p1 VALUES LESS THAN (4096)
+	)`)
+
+	_, err := tk.Exec("INSERT INTO t_month_data_monitor VALUES (5000, '2019-04-04')")
 	c.Assert(table.ErrNoPartitionForGivenValue.Equal(err), IsTrue)
 }