@imtbkcat 2020-07-28
Here is a simple `LOAD DATA` benchmark result:

Before this fix:
```
range partition table:

MySQL [test]> LOAD DATA LOCAL INFILE '2017Q3-capitalbikeshare-tripdata.csv' INTO TABLE trips_range FIELDS TERMINATED BY ',' ENCLOSED BY '\"' LINES TERMINATED BY '\r\n' IGNORE 1 LINES (duration, start_date, end_date, start_station_number, start_station, end_station_number, end_station, bike_number, member_type);
Query OK, 1191585 rows affected (25.168 sec)
Records: 1191585  Deleted: 0  Skipped: 0  Warnings: 0

hash partition table:

MySQL [test]> LOAD DATA LOCAL INFILE '2017Q3-capitalbikeshare-tripdata.csv' INTO TABLE trips_hash FIELDS TERMINATED BY ',' ENCLOSED BY '\"' LINES TERMINATED BY '\r\n' IGNORE 1 LINES (duration, start_date, end_date, start_station_number, start_station, end_station_number, end_station, bike_number, member_type);
Query OK, 1191585 rows affected (15.620 sec)
Records: 1191585  Deleted: 0  Skipped: 0  Warnings: 0

general table:

MySQL [test]> LOAD DATA LOCAL INFILE '2017Q3-capitalbikeshare-tripdata.csv' INTO TABLE trips_general FIELDS TERMINATED BY ',' ENCLOSED BY '\"' LINES TERMINATED BY '\r\n' IGNORE 1 LINES (duration, start_date, end_date, start_station_number, start_station, end_station_number, end_station, bike_number, member_type);
Query OK, 1191585 rows affected (9.432 sec)
Records: 1191585  Deleted: 0  Skipped: 0  Warnings: 0
```

After using optimize:
```
range:
MySQL [test]> LOAD DATA LOCAL INFILE '2017Q3-capitalbikeshare-tripdata.csv' INTO TABLE trips_range FIELDS TERMINATED BY ',' ENCLOSED BY '\"' LINES TERMINATED BY '\r\n' IGNORE 1 LINES (duration, start_date, end_date, start_station_number, start_station, end_station_number, end_station, bike_number, member_type);
Query OK, 1191585 rows affected (9.566 sec)
Records: 1191585  Deleted: 0  Skipped: 0  Warnings: 0

hash:
MySQL [test]> LOAD DATA LOCAL INFILE '2017Q3-capitalbikeshare-tripdata.csv' INTO TABLE trips_hash FIELDS TERMINATED BY ',' ENCLOSED BY '\"' LINES TERMINATED BY '\r\n' IGNORE 1 LINES (duration, start_date, end_date, start_station_number, start_station, end_station_number, end_station, bike_number, member_type);
Query OK, 1191585 rows affected (10.437 sec)
Records: 1191585  Deleted: 0  Skipped: 0  Warnings: 0
```

The performance improvement is very clear in `LOAD DATA` cases.