@tiancaiamao 2020-07-29
Please add more test to cover the new code.
Especially the +-1 case and null value.